hunting the bug

+++ pixray.py
@@ -165,7 +165,8 @@ class MakeCutouts(nn.Module):
         augmentations = []
         if global_aspect_width == 1:
             n_s = 0.95
-            n_t = (1-n_s)/2
+            # RS EDIT:
+            n_t = int((1-n_s)/2)
             augmentations.append(K.RandomAffine(degrees=0, translate=(n_t, n_t), scale=(n_s, n_s), p=1.0, return_transform=True))
         elif global_aspect_width > 1:
             n_s = 1/global_aspect_width
